refactor authoring and delivery to use a common part layout system
